Begun development on WitchFS

+++ WitchOS/Kernel API/Kernel_Memory.cs
@@ -10,41 +10,41 @@ namespace WitchOS
         {
             Cosmos.Core.Heap.MemAlloc(length);
         }
-        public static unsafe void MemRemove(byte start, uint offset, uint length)
+        public static unsafe void MemRemove(uint start, uint offset, uint length)
         {
             if (offset >= length) return;
-            byte* ptr = (byte*)start;
+            uint* ptr = (uint*)start;
             for (uint i = offset; i < offset + length; i++)
             {
-                ptr[i] = (byte)0;
+                ptr[i] = (uint)0;
             }
         }
-        public static unsafe void MemCopy(byte source, byte destination, uint offset, uint length)
+        public static unsafe void MemCopy(uint source, uint destination, uint offset, uint length)
         {
             if (offset >= length) return;
-            byte* src = (byte*)source;
-            byte* dst = (byte*)destination;
+            uint* src = (uint*)source;
+            uint* dst = (uint*)destination;
             for (uint i = offset; i < offset + length; i++)
             {
                 dst[i] = src[i];
             }
         }
-        public static unsafe void MemMove(byte source, byte destination, uint offset, uint length)
+        public static unsafe void MemMove(uint source, uint destination, uint offset, uint length)
         {
             if (offset >= length) return;
-            byte* src = (byte*)source;
-            byte* dst = (byte*)destination;
+            uint* src = (uint*)source;
+            uint* dst = (uint*)destination;
             for (uint i = offset; i < offset + length; i++)
             {
                 dst[i] = src[i];
                 src[i] = 0;
             }
         }
-        public static unsafe bool MemCompare(byte source1, byte source2, uint offset, uint length)
+        public static unsafe bool MemCompare(uint source1, uint source2, uint offset, uint length)
         {
             if (offset >= length) return false;
-            byte* ptr1 = (byte*)source1;
-            byte* ptr2 = (byte*)source2;
+            uint* ptr1 = (uint*)source1;
+            uint* ptr2 = (uint*)source2;
             for (uint i = offset; i < offset + length; i++)
             {
                 if (ptr1[i] != ptr2[i]) return false;
